+++ src/main/java/com/example/planergram/controller/StayLikeController.java
@@ -1,6 +1,5 @@
 package com.example.planergram.controller;
 
-import com.example.planergram.DTO.StayDTO;
 import com.example.planergram.DTO.StayLikeDTO;
 import com.example.planergram.service.StayLikeService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,23 +15,23 @@ public class StayLikeController {
 
     @GetMapping("/staylike")
     public String clickStayLike(
-            @RequestParam(value="user", defaultValue="0") int userId,
-            @RequestParam(value="stay", defaultValue="0") int stayId){
+            @RequestParam(value="user", defaultValue="0") Long userId,
+            @RequestParam(value="stay", defaultValue="0") Long stayId){
         return stayLikeService.clickStayLike(userId,stayId);
     }
 
     @GetMapping("/staylike/user/{userId}")
-    public List<StayLikeDTO> StayLikeFindByUser(@PathVariable int userId){
+    public List<StayLikeDTO> StayLikeFindByUser(@PathVariable Long userId){
         return stayLikeService.findByUser(userId);
     }
 
     @GetMapping("/staylike/stay/{stayId}")
-    public List<StayLikeDTO> StayLikeFindByStay(@PathVariable int stayId){
+    public List<StayLikeDTO> StayLikeFindByStay(@PathVariable Long stayId){
         return stayLikeService.findByStay(stayId);
     }
 
     @GetMapping("/staylike/{id}")
-    public StayLikeDTO findById(@PathVariable int id){
+    public StayLikeDTO findById(@PathVariable Long id){
         return stayLikeService.findById(id);
     }
 }
