+++ src/main/java/com/example/planergram/repository/StayLikeRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface StayLikeRepository extends JpaRepository<StayLike,Integer> {
+public interface StayLikeRepository extends JpaRepository<StayLike,Long> {
     StayLike findByUserAndStay(User user, Stay stay);
     List<StayLike> findByUser(User user);
     List<StayLike> findByStay(Stay stay);
