+++ src/main/java/com/example/planergram/repository/StayRepository.java
@@ -3,5 +3,5 @@ package com.example.planergram.repository;
 import com.example.planergram.model.Stay;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StayRepository extends JpaRepository<Stay,Integer> {
+public interface StayRepository extends JpaRepository<Stay,Long> {
 }
